+++ src/c_cmds.c
@@ -306,6 +306,7 @@ static void r_gamma_cvar_func2(char *cmd, char *parms);
 static void r_hud_cvar_func2(char *cmd, char *parms);
 static void r_hud_translucency_cvar_func2(char *cmd, char *parms);
 static void r_lowpixelsize_cvar_func2(char *cmd, char *parms);
+static void r_messagepos_cvar_func2(char *cmd, char *parms);
 static dboolean r_messagescale_cvar_func1(char *cmd, char *parms);
 static void r_messagescale_cvar_func2(char *cmd, char *parms);
 static void r_screensize_cvar_func2(char *cmd, char *parms);
@@ -624,6 +625,8 @@ consolecmd_t consolecmds[] =
         "Toggles the swirl effect of liquid sectors."),
     CVAR_SIZE(r_lowpixelsize, "", null_func1, r_lowpixelsize_cvar_func2,
         "The size of pixels when the graphic detail is low\n(<i>width</i><b>\xD7</b><i>height</i>)."),
+    CVAR_POS(r_messagepos, "", null_func1, r_messagepos_cvar_func2,
+        "The position of player messages (<b>(</b><i>x</i><b>,</b><i>y</i><b>)</b>)."),
     CVAR_BOOL(r_messagescale, "", r_messagescale_cvar_func1, r_messagescale_cvar_func2, SCALEVALUEALIAS,
         "The scale of player messages (<b>big</b> or <b>small</b>)."),
     CVAR_BOOL(r_mirroredweapons, "", bool_cvars_func1, bool_cvars_func2, BOOLVALUEALIAS,
@@ -4260,6 +4263,32 @@ static void gp_sensitivity_cvar_func2(char *cmd, char *parms)
 
     if (gp_sensitivity != gp_sensitivity_old)
         I_SetGamepadSensitivity(gp_sensitivity);
+}
+
+//
+// r_messagepos CVAR
+//
+static void r_messagepos_cvar_func2(char *cmd, char *parms)
+{
+    if (*parms)
+    {
+        if (!M_StringCompare(r_messagepos, parms))
+        {
+            r_messagepos = strdup(parms);
+            HU_GetMessagePosition();
+            M_SaveCVARs();
+        }
+    }
+    else
+    {
+        C_Output(removenewlines(consolecmds[C_GetIndex(stringize(r_messagepos))].description));
+
+        if (M_StringCompare(r_messagepos, r_messagepos_default))
+            C_Output("It is currently set to its default of <b>%s</b>.", r_messagepos);
+        else
+            C_Output("It is currently set to <b>%s</b> and its default is <b>%s</b>.",
+                r_messagepos, r_messagepos_default);
+    }
 }
 
 //